Begin refactoring into smaller files.
