+++ linkset/src/types/normalization.ts
@@ -1,0 +1,12 @@
+import { TargetAttributes } from './link';
+
+type TargetObject = { href: string } & TargetAttributes;
+
+type LinkSet = {
+  linkset: {
+    anchor: string;
+    [rel: string]: TargetObject[];
+  }[];
+};
+
+export default LinkSet;
